+++ Source/utils.h
@@ -3,7 +3,7 @@
 
 #include "structs.h"
 
-double clamp(double d, double min, double max);
+uint64_t clamp(uint64_t d, uint64_t min, uint64_t max);
 uint64_t get_time_ms(void);
 int input_waiting(void);
 void read_input(thread_t *thread);
